issue with message opt

+++ API/SignalR/MessageHub.cs
@@ -29,17 +29,19 @@ public class MessageHub(
         var groupName = GetGroupName(Context.User.GetUserName(), otherUser);
 
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
         var messages = await messageRepository.GetMessageThread(Context.User.GetUserName(), otherUser);
 
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
-
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -104,7 +106,7 @@ public class MessageHub(
         }
     }
 
-    private async Task<bool> AddToGroup(string groupName)
+    private async Task<Group> AddToGroup(string groupName)
     {
         var userName = Context.User?.GetUserName() ?? throw new HubException("cannot get username");
         var group = await messageRepository.GetMessageGroup(groupName);
@@ -118,18 +120,30 @@ public class MessageHub(
 
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync())
+        {
+            return group;
+        }
+
+        throw new HubException("Failed to join group");
     }
 
-    private async Task RemoveFromMessageGroup()
+    private async Task<Group> RemoveFromMessageGroup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
 
-        if (connection is not null)
+        if (connection is not null && group is not null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+
+            if (await messageRepository.SaveAllAsync())
+            {
+                return group;
+            }
         }
+
+        throw new HubException("Failed to remove from group");
     }
 
     private string GetGroupName(string sender, string receiver)
